requirements.txt
discord.py
asyncpg
aiofiles
aiointeractions
jishaku; python_version < "3.11"
matplotlib
